requirements.txt
python-dotenv==1.0.1
zeep==4.3.1
flask==3.1.0